+++ src/application/GridworldDRLExperienceReplayPOMDP.cpp
@@ -162,8 +162,8 @@ std::string GridworldDRLExperienceReplayPOMDP::streamNetworkResponseTable() {
 	mic::types::Position2D p = grid_env.getAgentPosition();
 
 	// Copy data.
-	for (long oy=0, ey=(p.y-dy); oy<grid_env.getObservationHeight(); oy++, ey++){
-		for (long ox=0, ex=(p.x-dx); ox<grid_env.getObservationWidth(); ox++, ex++) {
+	for (long oy=0, ey=(p.y-dy); oy<(long)grid_env.getObservationHeight(); oy++, ey++){
+		for (long ox=0, ex=(p.x-dx); ox<(long)grid_env.getObservationWidth(); ox++, ex++) {
 
 	//for (size_t y=0; y<grid_env.getObservationHeight(); y++){
 //		for (size_t x=0; x<grid_env.getObservationWidth(); x++) {
@@ -187,10 +187,10 @@ std::string GridworldDRLExperienceReplayPOMDP::streamNetworkResponseTable() {
 	rewards_table += "Action values:\n";
 	actions_table += "Best actions:\n";
 	// Generate all possible states and all possible rewards.
-	for (long oy=0, ey=(p.y-dy); oy<grid_env.getObservationHeight(); oy++, ey++){
+	for (long oy=0, ey=(p.y-dy); oy<(long)grid_env.getObservationHeight(); oy++, ey++){
 		rewards_table += "| ";
 		actions_table += "| ";
-		for (long ox=0, ex=(p.x-dx); ox<grid_env.getObservationWidth(); ox++, ex++) {
+		for (long ox=0, ex=(p.x-dx); ox<(long)grid_env.getObservationWidth(); ox++, ex++) {
 			float bestqval = -std::numeric_limits<float>::infinity();
 			size_t best_action = -1;
 			for (size_t a=0; a<4; a++) {
@@ -490,7 +490,7 @@ bool GridworldDRLExperienceReplayPOMDP::performSingleStep() {
 		return false;
 
 	// Check whether we reached maximum number of iterations.
-	if ((step_limit>0) && (iteration >= step_limit))
+	if ((step_limit>0) && (iteration >= (size_t)step_limit))
 		return false;
